Seperate list diffs from dict diffs
